requirements.txt
# CORE DEPENDENCIES
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# CONFIGURATION
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# DEEP LEARNING
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# OPENAI
openai>=1.0.0

# REDIS (EVENT-DRIVEN)
redis>=4.5.0

# GRPC
grpcio>=1.50.0
grpcio-tools>=1.50.0

# EVALUATION
rouge-score>=0.1.2
bert-score>=0.3.13

# VISUALIZATION 
seaborn>=0.12.0
matplotlib>=3.7.0
